+++ RELEASE-NOTES.md
@@ -2,10 +2,10 @@
 
 ## v2018.08.1
 
-- Add "Citizen Enrollment" feature.
 - Print Dead Person Record: change the record's length from 20 to 15 digits.
 - Support single fingerprint devices.
 - Show device controls if needed only, based on user roles.
+- Convicted Report: add 2 more crime types and change the gui for choosing to be dynamic.
 
 ## v2018.07.4